==> Norway_Morsa/MetData_Processing/notebooks/02_tidy_file_names.ipynb <==
import os
import shutil
import glob

# Tidy file names

Simplify file names and directory structure from climate4R scripts.

base_path = r'/home/jovyan/projects/watexr/data/Morsa/CLIMATE'

for season in ['winter', 'spring', 'early-summer', 'late-summer']:
    for member in range(1,16):
        dir_name = list(set(os.listdir(os.path.join(base_path, season))))[0]
        date_range = dir_name.split('_')[-1]
    
        fpath = os.path.join(base_path, 
                             season, 
                             f'Morsa_NIVA_System4_seasonal_15_seasonal_member{member:02d}_day_{date_range}',
                              'meteo_file.dat')
        
        new_fpath = os.path.join(base_path, 
                                 f's4_m{member:02d}_{season}_{date_range}.dat')
        shutil.copyfile(fpath, new_fpath)
