==> Norway_Morsa/MetData_Processing/notebooks/10_check_bc_issues.ipynb <==
%matplotlib inline
import pandas as pd
import matplotlib.pyplot as plt
plt.style.use("ggplot")

# Chech bias correction issues

Francois has notices some strage patterns in the bias corrected met data for GOTM. These issues only apply in autumn to the `rlds` variable (and only for GOTM). See Slack message from Francois received 15.09.2020 for more details.

The code below produces plots that illustrate the problem.

season = 'autumn'
model = 'gotm'

for member in range(1, 26):
    fpath = f'../../Data/Meteorological/07_s5_seasonal/s5_morsa_{model}_merged_{season}_member{member:02d}.csv'
    fpath_bc = f'../../Data/Meteorological/07_s5_seasonal/s5_morsa_{model}_merged_{season}_member{member:02d}_bc.csv'

    df = pd.read_csv(fpath)
    df_bc = pd.read_csv(fpath_bc)

    pars = list(df.columns)
    pars.remove('dates')

    df = pd.merge(df, df_bc, how='inner', on='dates', suffixes=['', '_bc'])

    fig, axes = plt.subplots(nrows=3, ncols=3, figsize=(10,10))
    axes = axes.flatten()

    for idx, par in enumerate(pars):
        axes[idx].plot(df[par], df[par + '_bc'], 'ro')
        axes[idx].plot(df[par], df[par], 'k-')
        axes[idx].set_xlabel('Uncorrected')
        axes[idx].set_ylabel('Bias corrected')
        axes[idx].set_title(par)

    plt.tight_layout()
    plt.savefig(f'./plots/bc_issues/member_{member:02d}.png')
    plt.close()
